Move heartbeat to separate genserver
